requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
Django==2.1.3
django-pyodbc-azure==2.1.0.0
django-rest-framework==0.1.0
djangorestframework==3.9.0
pyodbc==4.0.24
pytz==2018.7
